refactor(functions): extract build move
